Improve and unify some UT
